# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/classes/{id}": {
      "get": {
        "operationId": "ClassesController_get_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Class with course, teachers and roster counts",
        "tags": [
          "Classes"
        ]
      },
      "patch": {
        "operationId": "ClassesController_update_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClassDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update class details (assigned teachers and administrators)",
        "tags": [
          "Classes"
        ]
      },
      "delete": {
        "operationId": "ClassesController_remove_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel and soft-delete a class",
        "tags": [
          "Classes"
        ]
      }
    },
...
    "/api/v1/classes/{classId}/grading": {
      "get": {
        "operationId": "ClassGradingController_grading_api/v1",
        "parameters": [
          {
            "name": "classId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Grading settings of a class: mode, categories, late policy and syllabus (members and staff)",
        "tags": [
          "Grades"
        ]
      },
      "put": {
        "operationId": "ClassGradingController_setGrading_api/v1",
        "parameters": [
          {
            "name": "classId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClassGradingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Grades"
        ]
      }
    },
    "/api/v1/classes/{classId}/grading/categories": {
      "post": {
        "operationId": "ClassGradingController_createCategory_api/v1",
        "parameters": [
          {
            "name": "classId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Grades"
        ]
      }
    },
    "/api/v1/classes/{classId}/grading/categories/{id}": {
      "patch": {
        "operationId": "ClassGradingController_updateCategory_api/v1",
        "parameters": [
          {
            "name": "classId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Grades"
        ]
      },
      "delete": {
        "operationId": "ClassGradingController_removeCategory_api/v1",
        "parameters": [
          {
            "name": "classId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Grades"
        ]
      }
    },
    "/api/v1/classes/{classId}/gradebook": {
      "get": {
        "operationId": "ClassGradingController_book_api/v1",
        "parameters": [
          {
            "name": "classId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Gradebook: weighted categories, drop-lowest, extra credit, marks; standards levels for standards-based classes",
        "tags": [
          "Grades"
        ]
      }
    },
    "/api/v1/classes/{classId}/gradebook/export": {
      "get": {
        "operationId": "ClassGradingController_exportCsv_api/v1",
        "parameters": [
          {
            "name": "classId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Gradebook as CSV",
        "tags": [
          "Grades"
        ]
      }
    },
    "/api/v1/assignments/{assignmentId}/marks/{studentId}": {
      "put": {
        "operationId": "MarksController_set_api/v1",
        "parameters": [
          {
            "name": "assignmentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "studentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetMarkDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Set a mark for one student: missing, excused, incomplete, or null to clear",
        "tags": [
          "Grades"
        ]
      }
    },
    "/api/v1/organizations/{organizationId}/proficiency-scales": {
      "get": {
        "operationId": "ProficiencyScalesController_list_api/v1",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Grades"
        ]
      },
      "post": {
        "operationId": "ProficiencyScalesController_create_api/v1",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScaleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Grades"
        ]
      }
    },
    "/api/v1/organizations/{organizationId}/proficiency-scales/{id}": {
      "patch": {
        "operationId": "ProficiencyScalesController_update_api/v1",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScaleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Grades"
        ]
      },
      "delete": {
        "operationId": "ProficiencyScalesController_remove_api/v1",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Grades"
        ]
      }
    },
...
  },
  "info": {
    "title": "SmartSchool API",
    "description": "LMS API for SmartSchool. Conventions: docs/09-API-DESIGN.md. Errors use RFC 9457 problem details.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateClassDto": {
        "type": "object",
        "properties": {
          "courseId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Algebra I - Section A"
          },
          "section": {
            "type": "string",
            "example": "A"
          },
          "academicYearId": {
            "type": "string",
            "format": "uuid"
          },
          "termId": {
            "type": "string",
            "format": "uuid",
            "description": "Term from the school structure; fills the term label"
          },
          "periodId": {
            "type": "string",
            "format": "uuid",
            "description": "Period of the bell schedule"
          },
          "gradeLevel": {
            "type": "string",
            "example": "7"
          },
          "term": {
            "type": "string",
            "example": "2026-Fall",
            "description": "Free-text term label when no termId is given"
          },
          "startDate": {
            "type": "string",
            "example": "2026-09-01"
          },
          "endDate": {
            "type": "string",
            "example": "2026-12-18"
          },
          "room": {
            "type": "string"
          },
          "meetingSchedule": {
            "type": "string",
            "example": "Mon/Wed/Fri 09:00-09:50"
          },
          "maxStudents": {
            "type": "number",
            "example": 25
          },
          "status": {
            "type": "string",
            "enum": [
              "scheduled",
              "in_progress",
              "completed",
              "cancelled"
            ],
            "default": "scheduled"
          },
          "teacherId": {
            "type": "string",
            "format": "uuid",
            "description": "Primary teacher to assign on creation"
          },
          "organizationId": {
            "type": "string",
            "format": "uuid",
            "description": "District roles choose the organisation"
          }
        }
      },
...
      "ClassGradingDto": {
        "type": "object",
        "properties": {
          "gradingMode": {
            "type": "string",
            "enum": [
              "points",
              "standards"
            ]
          },
          "proficiencyScaleId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "latePolicy": {
            "type": "string",
            "description": "Shown to students and parents on the class page"
          },
          "syllabus": {
            "type": "string"
          }
        }
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Homework"
          },
          "weight": {
            "type": "number",
            "example": 20,
            "description": "Percent of the final grade"
          },
          "dropLowest": {
            "type": "number",
            "default": 0
          },
          "sortOrder": {
            "type": "number"
          }
        },
        "required": [
          "name",
          "weight"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Homework"
          },
          "weight": {
            "type": "number",
            "example": 20,
            "description": "Percent of the final grade"
          },
          "dropLowest": {
            "type": "number",
            "default": 0
          },
          "sortOrder": {
            "type": "number"
          }
        }
      },
      "SetMarkDto": {
        "type": "object",
        "properties": {
          "mark": {
            "type": "string",
            "enum": [
              "missing",
              "excused",
              "incomplete"
            ],
            "nullable": true,
            "description": "null clears the mark"
          },
          "note": {
            "type": "string"
          }
        },
        "required": [
          "mark"
        ]
      },
      "ProficiencyLevelDto": {
        "type": "object",
        "properties": {
          "level": {
            "type": "number",
            "example": 3
          },
          "label": {
            "type": "string",
            "example": "Proficient"
          },
          "minPercent": {
            "type": "number",
            "example": 80,
            "description": "Lowest percentage that maps to this level"
          }
        },
        "required": [
          "level",
          "label",
          "minPercent"
        ]
      },
      "CreateScaleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Four levels"
          },
          "levels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProficiencyLevelDto"
            }
          },
          "isDefault": {
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "levels"
        ]
      },
      "UpdateScaleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Four levels"
          },
          "levels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProficiencyLevelDto"
            }
          },
          "isDefault": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}